+++ apps/Standalone/src/designer/app/AzureLogicAppsDesigner/laDesigner.tsx
@@ -249,7 +249,24 @@ const DesignerEditor = () => {
               Default_Agent: {
                 type: 'Agent',
                 limit: {},
-                inputs: {},
+                inputs: {
+                  parameters: {
+                    deploymentId: '',
+                    messages: '',
+                    agentModelType: 'AzureOpenAI',
+                    agentModelSettings: {
+                      agentHistoryReductionSettings: {
+                        agentHistoryReductionType: 'maximumTokenCountReduction',
+                        maximumTokenCount: 128000,
+                      },
+                    },
+                  },
+                  modelConfigurations: {
+                    model1: {
+                      referenceName: '',
+                    },
+                  },
+                },
                 tools: {},
                 runAfter: {},
               },
@@ -264,7 +281,24 @@ const DesignerEditor = () => {
             Default_Agent: {
               type: 'Agent',
               limit: {},
-              inputs: {},
+              inputs: {
+                parameters: {
+                  deploymentId: '',
+                  messages: '',
+                  agentModelType: 'AzureOpenAI',
+                  agentModelSettings: {
+                    agentHistoryReductionSettings: {
+                      agentHistoryReductionType: 'maximumTokenCountReduction',
+                      maximumTokenCount: 128000,
+                    },
+                  },
+                },
+                modelConfigurations: {
+                  model1: {
+                    referenceName: '',
+                  },
+                },
+              },
               tools: {},
               runAfter: {},
             },
@@ -524,6 +558,7 @@ const DesignerEditor = () => {
                   }}
                   switchViews={handleSwitchView}
                   saveWorkflowFromCode={saveWorkflowFromCode}
+                  setWorkflow={setWorkflow}
                 />
                 {designerView ? <Designer /> : <CodeViewEditor ref={codeEditorRef} workflowKind={workflow?.kind} />}
                 <CombineInitializeVariableDialog />