+++ libs/designer/src/lib/core/queries/runs.ts
@@ -1,4 +1,4 @@
-import { useQuery } from '@tanstack/react-query';
+import { useMutation, useQuery } from '@tanstack/react-query';
 import { isNullOrUndefined, type LogicAppsV2, type Run, RunService } from '@microsoft/logic-apps-shared';
 import { getReactQueryClient } from '../ReactQueryProvider';
 import { isRunError } from '@microsoft/designer-ui';
@@ -25,6 +25,8 @@ export const runsQueriesKeys = {
   useAgentActionsRepetition: 'useAgentActionsRepetition',
   useChatHistory: 'useChatHistory',
   useAgentChatInvokeUri: 'useAgentChatInvokeUri',
+  useRunInstance: 'useRunInstance',
+  useCancelRun: 'useCancelRun',
 };
 
 export const useRuns = (enabled = false) => {
@@ -165,6 +167,12 @@ export const useAgentRepetition = (
       enabled: isMonitoringView && runIndex !== undefined && isAgent,
     }
   );
+};
+
+export const useCancelRun = (runId: string) => {
+  return useMutation([runsQueriesKeys.useCancelRun, { runId }], async () => {
+    return await RunService().cancelRun(runId);
+  });
 };
 
 export const useAgentActionsRepetition = (