+++ libs/designer/src/lib/core/templates/utils/helper.ts
@@ -264,8 +264,8 @@ export const checkWorkflowNameWithRegex = (intl: IntlShape, workflowName: string
   const regex = /^[A-Za-z][A-Za-z0-9]*(?:[_-][A-Za-z0-9]+)*$/;
   if (!regex.test(workflowName)) {
     return intl.formatMessage({
-      defaultMessage: 'Name does not match the given pattern.',
-      id: 'zMKxg9',
+      defaultMessage: "Name can only contain letters, numbers, and '-', '(', ')', '_' or '.",
+      id: 'TwBRcg',
       description: 'Error message when the workflow name is invalid regex.',
     });
   }