+++ libs/designer/src/lib/ui/configuretemplate/tabs/useWizardTabs.tsx
@@ -9,7 +9,6 @@ import { publishTab } from './publishTab';
 import { reviewTab } from './reviewTab';
 import { useTemplatesStrings } from '../../templates/templatesStrings';
 import { useResourceStrings } from '../resources';
-import { useEffect } from 'react';
 import { setRunValidation } from '../../../core/state/templates/tabSlice';
 import {
   validateParameterDetails,
@@ -17,6 +16,10 @@ import {
   validateWorkflowManifestsData,
 } from '../../../core/state/templates/templateSlice';
 import constants from '../../../common/constants';
+import type { Template } from '@microsoft/logic-apps-shared';
+import { TemplateResourceService } from '@microsoft/logic-apps-shared';
+import { useQueryClient } from '@tanstack/react-query';
+import { useEffect, useCallback } from 'react';
 
 export const useConfigureTemplateWizardTabs = ({
   onSaveWorkflows,
@@ -27,18 +30,30 @@ export const useConfigureTemplateWizardTabs = ({
 }) => {
   const intl = useIntl();
   const dispatch = useDispatch<AppDispatch>();
+  const queryClient = useQueryClient();
   const resources = { ...useTemplatesStrings().tabLabelStrings, ...useResourceStrings() };
 
-  const { selectedTabId, enableWizard, isWizardUpdating, workflows, parametersHasError, templateManifestHasError, runValidation } =
-    useSelector((state: RootState) => ({
-      selectedTabId: state.tab.selectedTabId,
-      enableWizard: state.tab.enableWizard,
-      isWizardUpdating: state.tab.isWizardUpdating,
-      runValidation: state.tab.runValidation,
-      workflows: state.template.workflows,
-      parametersHasError: Object.values(state.template.errors.parameters).some((value) => value !== undefined),
-      templateManifestHasError: Object.values(state.template.errors.manifest).some((value) => value !== undefined),
-    }));
+  const {
+    enableWizard,
+    isWizardUpdating,
+    templateManifest,
+    state,
+    workflows,
+    parametersHasError,
+    templateManifestHasError,
+    runValidation,
+    selectedTabId,
+  } = useSelector((state: RootState) => ({
+    enableWizard: state.tab.enableWizard,
+    isWizardUpdating: state.tab.isWizardUpdating,
+    runValidation: state.tab.runValidation,
+    templateManifest: state.template.manifest,
+    state: state.template.status,
+    workflows: state.template.workflows,
+    parametersHasError: Object.values(state.template.errors.parameters).some((value) => value !== undefined),
+    templateManifestHasError: Object.values(state.template.errors.manifest).some((value) => value !== undefined),
+    selectedTabId: state.tab.selectedTabId,
+  }));
 
   const hasAnyWorkflowErrors = Object.values(workflows).some(
     ({ errors }) =>
@@ -59,9 +74,30 @@ export const useConfigureTemplateWizardTabs = ({
     }
   }, [dispatch, selectedTabId]);
 
+  const onTemplateSave = useCallback(
+    async (publishState: string) => {
+      const manifestToUpdate: Template.TemplateManifest = {
+        ...(templateManifest as Template.TemplateManifest),
+        details: {
+          ...templateManifest?.details,
+          Type: Object.keys(workflows).length > 1 ? 'Accelerator' : 'Workflow',
+        } as any,
+      };
+      const templateId = templateManifest?.id as string;
+      await TemplateResourceService().updateTemplate(templateId, manifestToUpdate, publishState);
+
+      queryClient.removeQueries(['template', templateId.toLowerCase()]);
+    },
+    [queryClient, templateManifest, workflows]
+  );
+
+  const onProfileSave = useCallback(async () => onTemplateSave(state as string), [onTemplateSave, state]);
+  const disableProfileSave = false;
+
   return [
     workflowsTab(resources, dispatch, onSaveWorkflows, {
       tabStatusIcon: hasAnyWorkflowErrors ? 'error' : runValidation ? 'success' : 'in-progress',
+      disabled: !enableWizard || isWizardUpdating,
     }),
     connectionsTab(intl, resources, dispatch, {
       tabStatusIcon: enableWizard ? 'success' : undefined,
@@ -71,9 +107,11 @@ export const useConfigureTemplateWizardTabs = ({
       tabStatusIcon: parametersHasError ? 'error' : enableWizard ? (runValidation ? 'success' : 'in-progress') : undefined,
       disabled: !enableWizard || isWizardUpdating,
     }),
-    profileTab(resources, dispatch, {
+    profileTab(intl, resources, dispatch, {
       tabStatusIcon: templateManifestHasError ? 'error' : runValidation ? 'success' : enableWizard ? 'in-progress' : undefined,
       disabled: !enableWizard || isWizardUpdating,
+      onSave: onProfileSave,
+      disableSave: disableProfileSave,
     }),
     reviewTab(resources, dispatch, {
       tabStatusIcon: undefined,