+++ libs/designer/src/lib/ui/configuretemplate/workflows/workflowslist.tsx
@@ -23,7 +23,6 @@ import { useIntl } from 'react-intl';
 import { CommandBar, type ICommandBarItemProps } from '@fluentui/react';
 import { useCallback, useMemo } from 'react';
 import { openPanelView, TemplatePanelView } from '../../../core/state/templates/panelSlice';
-import { ConfigureWorkflowsPanel } from '../../panel/configureTemplatePanel/configureWorkflowsPanel/configureWorkflowsPanel';
 import { useFunctionalState } from '@react-hookz/web';
 import { Add12Filled } from '@fluentui/react-icons';
 import { deleteWorkflowData } from '../../../core/actions/bjsworkflow/configuretemplate';
@@ -31,6 +30,7 @@ import { useResourceStrings } from '../resources';
 import { useTemplatesStrings } from '../../templates/templatesStrings';
 import { WorkflowKind } from '../../../core/state/workflow/workflowInterfaces';
 import { equals } from '@microsoft/logic-apps-shared';
+import { ConfigureWorkflowsPanel } from '../panels/configureWorkflowsPanel/configureWorkflowsPanel';
 
 export const DisplayWorkflows = ({ onSave }: { onSave: (isMultiWorkflow: boolean) => void }) => {
   const intl = useIntl();
@@ -218,14 +218,14 @@ export const DisplayWorkflows = ({ onSave }: { onSave: (isMultiWorkflow: boolean
         }}
       />
 
-{workflowNamesWithErrors.length ? (
-      <MessageBar intent="error" className="msla-templates-error-message-bar">
-        <MessageBarBody>
-          <MessageBarTitle>{customResourceStrings.MissingRequiredFields}</MessageBarTitle>
-          <Text>{workflowNamesWithErrors.join(', ')}</Text>
-        </MessageBarBody>
-      </MessageBar>
-) : null}
+      {workflowNamesWithErrors.length ? (
+        <MessageBar intent="error" className="msla-templates-error-message-bar">
+          <MessageBarBody>
+            <MessageBarTitle>{customResourceStrings.MissingRequiredFields}</MessageBarTitle>
+            <Text>{workflowNamesWithErrors.join(', ')}</Text>
+          </MessageBarBody>
+        </MessageBar>
+      ) : null}
 
       {Object.keys(workflows).length > 0 ? (
         <Table aria-label={customResourceStrings.WorkflowsListTableLabel} style={{ minWidth: '550px' }}>